Add equals operator to Savepoint in Python API
After this commit the expression "sp1 == sp2" will return True for any
couple of savepoints whose name and metainfo match. Before this the
expression returned False if sp1 and sp2 were not the same object.

This also enables for checking if a copy of a savepoint can be found in
a list. For instance, if ser1 and ser2 are two independent serializers,
the following expression will now check if the first savepoint in ser1
can be found in ser2 as well:

    ser1.savepoints[0] in ser2.savepoints